backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
anthropic>=0.42.0
ffmpeg-python==0.2.0
pydantic>=2.9.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
websockets==13.0
python-dotenv==1.0.1
aiofiles==24.1.0
Pillow>=10.4.0
httpx[socks]>=0.27.0
greenlet>=3.0.0
